Update README.md with relative links to in-repo references
Changed links to use relative URIs instead of absolute URLs so that when the repo is forked it will remain self referential and not carry docs that are not referenced.  (There may be good reasons to have a singular documentation point, but it seems they may be obviated by the act of forking, itself. )

Signed-off-by: Meta <59613197+metagrapher@users.noreply.github.com>